# Optimize Kyber