--- a/.github/workflows/test_ci.yml
+++ b/.github/workflows/test_ci.yml
@@ -15,6 +15,8 @@ jobs:
     - uses: actions/checkout@v3
     - name: Setup Compiler
       run: sudo update-alternatives --install /usr/bin/g++ g++ /usr/bin/g++-10 10; sudo update-alternatives --install /usr/bin/g++ g++ /usr/bin/g++-9 9
+    - name: Fetch Dependency
+      run: git submodule update --init
     - name: Execute Tests
       run: make
     - name: Cleanup