--- a/Makefile
+++ b/Makefile
@@ -1,19 +1,12 @@
-CXX = g++
+CXX = clang++
 CXXFLAGS = -std=c++20 -Wall -Wextra -pedantic
 OPTFLAGS = -O3 -march=native
 IFLAGS = -I ./include
-DEP_IFLAGS = -I ./deps/sha3/include
+DEP_IFLAGS = -I ./sha3/include
 
 all: testing
 
-deps/sha3:
-	mkdir -p deps
-	git clone https://github.com/itzmeanjan/sha3.git $@
-	cd deps/sha3; git checkout 5886375; cd -
-
-dep: deps/sha3
-
-test/a.out: test/main.cpp include/*.hpp dep
+test/a.out: test/main.cpp include/*.hpp sha3/include/*.hpp
 	$(CXX) $(CXXFLAGS) $(OPTFLAGS) $(IFLAGS) $(DEP_IFLAGS) $< -o $@
 
 testing: test/a.out
@@ -26,7 +19,7 @@ clean:
 format:
 	find . -name '*.cpp' -o -name '*.hpp' | xargs clang-format -i --style=Mozilla
 
-bench/a.out: bench/main.cpp include/*.hpp dep
+bench/a.out: bench/main.cpp include/*.hpp sha3/include/*.hpp
 	# make sure you've google-benchmark globally installed;
 	# see https://github.com/google/benchmark/tree/60b16f1#installation
 	$(CXX) $(CXXFLAGS) $(OPTFLAGS) $(IFLAGS) $(DEP_IFLAGS) $< -lbenchmark -o $@