--- a/bench/main.cpp
+++ b/bench/main.cpp
@@ -4,8 +4,10 @@
 BENCHMARK(bench_kyber::ff_add);
 BENCHMARK(bench_kyber::ff_compound_add);
 BENCHMARK(bench_kyber::ff_sub);
+BENCHMARK(bench_kyber::ff_compound_sub);
 BENCHMARK(bench_kyber::ff_neg);
 BENCHMARK(bench_kyber::ff_mul);
+BENCHMARK(bench_kyber::ff_compound_mul);
 BENCHMARK(bench_kyber::ff_inv);
 BENCHMARK(bench_kyber::ff_div);
 BENCHMARK(bench_kyber::ff_exp);