--- a/include/bench_ff.hpp
+++ b/include/bench_ff.hpp
@@ -80,6 +80,31 @@ ff_sub(benchmark::State& state)
   assert(c == b);
 }
 
+// Benchmark compound subtraction over prime field Z_q | q = 3329
+void
+ff_compound_sub(benchmark::State& state)
+{
+  ff::ff_t a = ff::ff_t::random();
+  ff::ff_t b{};
+
+  for (auto _ : state) {
+    b -= a;
+
+    benchmark::DoNotOptimize(b);
+    benchmark::DoNotOptimize(a);
+    benchmark::ClobberMemory();
+  }
+
+  const size_t itr = state.iterations();
+
+  ff::ff_t c{};
+  for (size_t i = 0; i < itr; i++) {
+    c = c - a;
+  }
+
+  assert(c == b);
+}
+
 // Benchmark negation over prime field Z_q | q = 3329
 void
 ff_neg(benchmark::State& state)
@@ -108,6 +133,31 @@ ff_mul(benchmark::State& state)
 
   for (auto _ : state) {
     b = b * a;
+
+    benchmark::DoNotOptimize(b);
+    benchmark::DoNotOptimize(a);
+    benchmark::ClobberMemory();
+  }
+
+  const size_t itr = state.iterations();
+
+  ff::ff_t c{ 1 };
+  for (size_t i = 0; i < itr; i++) {
+    c = c * a;
+  }
+
+  assert(c == b);
+}
+
+// Benchmark compound multiplication over prime field Z_q | q = 3329
+void
+ff_compound_mul(benchmark::State& state)
+{
+  ff::ff_t a = ff::ff_t::random();
+  ff::ff_t b{ 1 };
+
+  for (auto _ : state) {
+    b *= a;
 
     benchmark::DoNotOptimize(b);
     benchmark::DoNotOptimize(a);