--- a/include/bench_kyber.hpp
+++ b/include/bench_kyber.hpp
@@ -5,3 +5,4 @@
 #include "bench_ntt.hpp"
 #include "bench_pke_kyber.hpp"
 #include "bench_serialize.hpp"
+#include "bench_utils.hpp"