--- a/include/bench_ntt.hpp
+++ b/include/bench_ntt.hpp
@@ -12,24 +12,28 @@ ntt(benchmark::State& state)
 {
   ff::ff_t poly_a[ntt::N]{};
   ff::ff_t poly_b[ntt::N]{};
-  ff::ff_t poly_c[ntt::N]{};
 
   for (size_t i = 0; i < ntt::N; i++) {
     poly_a[i] = ff::ff_t::random();
   }
 
-  for (auto _ : state) {
-    ntt::ntt(poly_a, poly_b);
+  std::memcpy(poly_b, poly_a, sizeof(ff::ff_t) * ntt::N);
 
-    benchmark::DoNotOptimize(poly_a);
+  for (auto _ : state) {
+    ntt::ntt(poly_b);
+
     benchmark::DoNotOptimize(poly_b);
     benchmark::ClobberMemory();
   }
 
-  ntt::intt(poly_b, poly_c);
+  const size_t itr = state.iterations();
+
+  for (size_t i = 0; i < itr; i++) {
+    ntt::ntt(poly_a);
+  }
 
   for (size_t i = 0; i < ntt::N; i++) {
-    assert(poly_a[i] == poly_c[i]);
+    assert(poly_a[i] == poly_b[i]);
   }
 }
 
@@ -40,24 +44,28 @@ intt(benchmark::State& state)
 {
   ff::ff_t poly_a[ntt::N]{};
   ff::ff_t poly_b[ntt::N]{};
-  ff::ff_t poly_c[ntt::N]{};
 
   for (size_t i = 0; i < ntt::N; i++) {
     poly_a[i] = ff::ff_t::random();
   }
 
-  ntt::ntt(poly_a, poly_b);
+  std::memcpy(poly_b, poly_a, sizeof(ff::ff_t) * ntt::N);
 
   for (auto _ : state) {
-    ntt::intt(poly_b, poly_c);
+    ntt::intt(poly_b);
 
     benchmark::DoNotOptimize(poly_b);
-    benchmark::DoNotOptimize(poly_c);
     benchmark::ClobberMemory();
   }
 
+  const size_t itr = state.iterations();
+
+  for (size_t i = 0; i < itr; i++) {
+    ntt::intt(poly_a);
+  }
+
   for (size_t i = 0; i < ntt::N; i++) {
-    assert(poly_a[i] == poly_c[i]);
+    assert(poly_a[i] == poly_b[i]);
   }
 }