--- a/include/decryption.hpp
+++ b/include/decryption.hpp
@@ -57,12 +57,9 @@ decrypt(
   }
 
   // step 4
-  ff::ff_t u_prime[k * ntt::N]{};
-
   for (size_t i = 0; i < k; i++) {
     const size_t uoff = i * ntt::N;
-
-    ntt::ntt(u + uoff, u_prime + uoff);
+    ntt::ntt(u + uoff);
   }
 
   ff::ff_t t[ntt::N]{};
@@ -73,15 +70,14 @@ decrypt(
   for (size_t i = 0; i < k; i++) {
     const size_t off = i * ntt::N;
 
-    ntt::polymul(s_prime + off, u_prime + off, tmp);
+    ntt::polymul(s_prime + off, u + off, tmp);
 
     for (size_t l = 0; l < ntt::N; l++) {
       t[l] += tmp[l];
     }
   }
 
-  ntt::intt(t, tmp);
-  std::memcpy(t, tmp, sizeof(tmp));
+  ntt::intt(t);
 
   for (size_t i = 0; i < ntt::N; i++) {
     v[i] = kyber_utils::compress<1>(v[i] - t[i]);