--- a/include/test_cpa_pke.hpp
+++ b/include/test_cpa_pke.hpp
@@ -40,8 +40,8 @@ test_kyber_cpa_pke()
   std::memset(enc, 0, enclen);
   std::memset(dec, 0, mlen);
 
-  random_data<uint8_t>(txt, mlen);
-  random_data<uint8_t>(rcoin, mlen);
+  kyber_utils::random_data<uint8_t>(txt, mlen);
+  kyber_utils::random_data<uint8_t>(rcoin, mlen);
 
   cpapke::keygen<k, eta1>(pkey, skey);
   cpapke::encrypt<k, eta1, eta2, du, dv>(pkey, txt, rcoin, enc);