- [x] Make (i)NTT implementation in-place -- avoiding unnecessary `memcpy`
- [x] Introduce compound subtraction and multiplication operators, when operating with `ff::ff_t` type

> Saving ~2 ns for subtraction and multiplication over Z_q

---

At commit dd47cb2419f1ca793b9fd57b284f1e1c97434297, on **Intel(R) Core(TM) i5-8279U CPU @ 2.40GHz**

```bash
2022-09-28T13:12:33+04:00
Running ./bench/a.out
Run on (8 X 2400 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB
  L1 Instruction 32 KiB
  L2 Unified 256 KiB (x4)
  L3 Unified 6144 KiB
Load Average: 2.08, 1.98, 1.91
-----------------------------------------------------------------------------------
Benchmark                                         Time             CPU   Iterations
-----------------------------------------------------------------------------------
bench_kyber::ff_add                            4.06 ns         4.05 ns    171697822
bench_kyber::ff_compound_add                   1.92 ns         1.92 ns    362560923
bench_kyber::ff_sub                            4.24 ns         4.23 ns    164150859
bench_kyber::ff_compound_sub                   2.10 ns         2.10 ns    333342857
bench_kyber::ff_neg                           0.828 ns        0.822 ns    854930506
bench_kyber::ff_mul                            6.74 ns         6.73 ns     97859669
bench_kyber::ff_compound_mul                   4.48 ns         4.47 ns    156239538
bench_kyber::ff_inv                            66.9 ns         66.9 ns      8785361
bench_kyber::ff_div                            76.1 ns         76.1 ns     11760051
bench_kyber::ff_exp                             410 ns          410 ns      1759156
bench_kyber::encode<1>                          148 ns          148 ns      4734560 bytes_per_second=206.057M/s
bench_kyber::decode<1>                          144 ns          144 ns      4870073 bytes_per_second=212.067M/s
bench_kyber::encode<4>                          927 ns          927 ns       738693 bytes_per_second=131.72M/s
bench_kyber::decode<4>                          900 ns          898 ns       777605 bytes_per_second=135.865M/s
bench_kyber::encode<5>                         1805 ns         1803 ns       390810 bytes_per_second=84.6427M/s
bench_kyber::decode<5>                         1653 ns         1651 ns       426892 bytes_per_second=92.4135M/s
bench_kyber::encode<10>                        2946 ns         2941 ns       238515 bytes_per_second=103.768M/s
bench_kyber::decode<10>                        3022 ns         3018 ns       233518 bytes_per_second=101.135M/s
bench_kyber::encode<11>                        4170 ns         4159 ns       168945 bytes_per_second=80.7085M/s
bench_kyber::decode<11>                        3870 ns         3866 ns       181117 bytes_per_second=86.8249M/s
bench_kyber::encode<12>                        3522 ns         3518 ns       198937 bytes_per_second=104.082M/s
bench_kyber::decode<12>                        3580 ns         3575 ns       197125 bytes_per_second=102.429M/s
bench_kyber::compress<1>                       5.23 ns         5.23 ns    130404814
bench_kyber::decompress<1>                     4.57 ns         4.57 ns    151764150
bench_kyber::compress<4>                       5.24 ns         5.24 ns    131467743
bench_kyber::decompress<4>                     4.84 ns         4.83 ns    143986112
bench_kyber::compress<5>                       5.23 ns         5.23 ns    131878897
bench_kyber::decompress<5>                     4.83 ns         4.82 ns    144529855
bench_kyber::compress<10>                      5.01 ns         5.00 ns    136681376
bench_kyber::decompress<10>                    4.83 ns         4.82 ns    145118925
bench_kyber::compress<11>                      5.07 ns         5.06 ns    138350858
bench_kyber::decompress<11>                    4.86 ns         4.86 ns    135054311
bench_kyber::ntt                               1615 ns         1592 ns       439329
bench_kyber::intt                              1508 ns         1497 ns       460608
bench_kyber::polymul                            367 ns          366 ns      1864211
bench_kyber::random_data<32>                  11618 ns        11459 ns        69520
bench_kyber::pke_keygen<2, 3>                 55224 ns        54716 ns        12662 items_per_second=18.2763k/s
bench_kyber::encrypt<2, 3, 2, 10, 4>          42178 ns        41758 ns        16641 items_per_second=23.9475k/s
bench_kyber::decrypt<2, 3, 2, 10, 4>          20202 ns        20113 ns        32982 items_per_second=49.7184k/s
bench_kyber::pke_keygen<3, 2>                 77117 ns        76626 ns         8946 items_per_second=13.0503k/s
bench_kyber::encrypt<3, 2, 2, 10, 4>          68981 ns        68248 ns        10120 items_per_second=14.6524k/s
bench_kyber::decrypt<3, 2, 2, 10, 4>          29876 ns        29624 ns        23459 items_per_second=33.7568k/s
bench_kyber::pke_keygen<4, 2>                131868 ns       128842 ns         6593 items_per_second=7.76145k/s
bench_kyber::encrypt<4, 2, 2, 11, 5>         118440 ns       114292 ns         6479 items_per_second=8.74952k/s
bench_kyber::decrypt<4, 2, 2, 11, 5>          42402 ns        42234 ns        15612 items_per_second=23.6774k/s
bench_kyber::kem_keygen<2, 3>                 63160 ns        63096 ns         9851 items_per_second=15.8488k/s
bench_kyber::encapsulate<2, 3, 2, 10, 4>      55426 ns        55368 ns        11332 items_per_second=18.061k/s
bench_kyber::decapsulate<2, 3, 2, 10, 4>      63287 ns        62948 ns        10263 items_per_second=15.8861k/s
bench_kyber::kem_keygen<3, 2>                 87242 ns        87143 ns         7413 items_per_second=11.4754k/s
bench_kyber::encapsulate<3, 2, 2, 10, 4>      82806 ns        82684 ns         8121 items_per_second=12.0942k/s
bench_kyber::decapsulate<3, 2, 2, 10, 4>      96234 ns        96157 ns         7160 items_per_second=10.3996k/s
bench_kyber::kem_keygen<4, 2>                120004 ns       119906 ns         5591 items_per_second=8.33987k/s
bench_kyber::encapsulate<4, 2, 2, 11, 5>     120743 ns       120661 ns         5266 items_per_second=8.28768k/s
bench_kyber::decapsulate<4, 2, 2, 11, 5>     148354 ns       148129 ns         4706 items_per_second=6.75087k/s
```